fix mobile/desktop toggle behavior
This change fixes the white background issue on desktop. The problem occurs due to the menu toggle being activated when a desktop user clicks on a result in the town selector. This change fixes this by detecting mobile devices using window.matchMedia, and only toggling the menu when the watchMedia is "only screen and (max-device-width: 480px)".